pyproject.toml
[build-system]
requires = ["maturin>=1.6,<2.0"]
build-backend = "maturin"

[project]
name = "laboneq"
version = "2.42.0"
description = "Zurich Instruments LabOne Q software framework for quantum computing control"
license = { text = "Apache 2.0" }
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Zurich Instruments Development Team", email = "info@zhinst.com" },
]
keywords = [
    "quantum",
    "sdk",
    "zhinst",
]
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "engineering_notation >= 0.10.0",
    "flask >= 3.0.2",
    "intervaltree >= 3.1.0",
    "jsonschema >= 4.21.1",
    "matplotlib >= 3.8.3",
    "cycler >= 0.12.1",
    "numpy >= 1.26.4, < 2.0",
    "openpulse ~= 1.0.1",
    "openqasm3 ~= 1.0.0",
    "orjson >= 3.9.15",
    "pybase64 >= 1.3.2",
    "pycparser >= 2.21",
    "python-box >= 7.1.1",
    "pyyaml >= 6.0.1",
    "requests >= 2.31.0",
    "rich >= 13.7.1",
    "scipy >= 1.12.0",
    "sqlitedict >= 2.1.0",
    "sortedcollections >= 2.1.0",
    "lagom >= 2.6.0",
    "attrs >= 20.1.0",
    "sortedcontainers >= 2.4.0",
    "typing_extensions >= 4.10.0",
    "zhinst-core ~= 24.10.65175",
    "zhinst-toolkit == 0.7.0",
    "zhinst-utils == 0.5.0",
    "zhinst-timing-models ~= 24.10.64896",
    "labone == 3.2.0",
    "unsync == 1.4.0",
    "pygments >= 2.18.0"
]

[project.urls]
Homepage = "https://github.com/zhinst/laboneq"

[tool.maturin]
manifest-path = "./src/rust/laboneq-rust/Cargo.toml"
module-name = "laboneq._rust"
features = []
python-source = "src/python"
include = [
    "laboneq/compiler/qccs-schema_2_5_0.json",
    "laboneq/pulse_sheet_viewer/pulse_sheet_viewer_template.html",
]
python-packages = ["laboneq"]
compatibility = "manylinux2014"

[tool.ruff]
line-length = 88
target-version = "py39"
exclude = [
    ".cargo",
    ".eggs",
    ".git",
    ".ruff_cache",
    ".tox",
    ".venv",
    "_build",
    "build",
    "dist",
    "target",
    "venv",
]

[tool.ruff.lint]
select = [
    # flake8-bugbear
    "B",
    # perflint
    "PERF",
    # pyflakes
    "F",
    # pycodestyle
    "E",
    "W",
    # flake8-2020
    "YTT",
    # flake8-quotes
    "Q",
    "B019",
    # isort
    "I002",
    # future annotations
    "FA"
]

ignore = [
    "E501",
    "F405",
    "E741",
    "E722",
    "E402",
    # Do not assign a lambda expression, use a def
    "E731"
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

mccabe.max-complexity = 10

[tool.ruff.lint.per-file-ignores]
# F401: Imported but unused
"__init__.py" = ["F401"]
"*/laboneq/simple.py" = ["F401"]

# Documentation notebooks (nbqa converts notebooks to .py)
"examples/**/*.ipynb" = [
  # Allow redefining values
  "F811",
  # Useless attribute access can display values in notebooks
  "B018",
]

# Ignore star import, mainly from laboneq.simple (TODO: Fix this)
"tests/integration/dsl_v3/*" = ["F403", "F405"]

"tests/*" = ["B008", "B011", "B017", "B904", "PERF"]
"resources/examples/*" = ["B015", "PERF"]
"examples/*" = ["PERF"]
"*.ipynb" = [
    "F403",  # Notebooks utilize `from laboneq.simple import *`
]

[tool.ruff.format]
docstring-code-format = true


[[tool.mypy.overrides]]
module = [
    "scipy.*",
    "sortedcollections",
    "sortedcontainers",
    "intervaltree",
    "pybase64",
    "engineering_notation",
    "jsonschema.*",
    "pycparser.*",
    "sqlitedict",
    "openpulse",
    "cycler"
]
ignore_missing_imports = true
